pyproject.toml
[tool.poetry]
name = "argoapichallenge"
version = "0.1.0"
description = "application suggested in a job interview by the company argo"
authors = ["Lucas Praciano <luskcct@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
dynaconf = "^3.1.12"
pytest = "^7.4.0"
uvicorn = "^0.22.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.18"}
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.6"
passlib = "^1.7.4"
pydantic = {version = "1.10.1", extras = ["email"]}
asyncpg = "^0.28.0"
bcrypt = "^4.0.1"
fastapi-mail = "^1.3.1"
pytz = "^2023.3"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
